+++ internal/install/blocky.go
@@ -73,8 +73,8 @@ blocking:
       - ads
   blockType: nxDomain
 ports:
-  dns: 53
-  http: 4000
+  dns: 127.0.0.1:53
+  http: 127.0.0.1:4000
 log:
   level: info
 `